# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Celebra Su Vida integration api",
        "description": "This Api is private. To use it, contact us to generate your credentials.",
        "contact": {
            "email": "admin@admin.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{schema}://albia.celebrasuvida.es/api",
            "description": "development server",
            "variables": {
                "schema": {
                    "enum": [
                        "https",
                        "http"
                    ],
                    "default": "https"
                }
            }
        },
        {
            "url": "{schema}://albia.local/api",
            "description": "development server local",
            "variables": {
                "schema": {
                    "enum": [
                        "https",
                        "http"
                    ],
                    "default": "http"
                }
            }
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "",
                "operationId": "auth/login",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jebauza@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "example": true
                                        },
                                        "message": {
                                            "example": "User login successfully."
                                        },
                                        "token_type": {
                                            "example": "Bearer"
                                        },
                                        "expires_at": {
                                            "example": "2022-03-02 17:10:15"
                                        },
                                        "access_token": {
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJhdWQiOiIxIiwianRpIjoiM2NhOGI5NzhiNDRmMDM5ODYyZDcyZDcwYmQ1ZDFlNGNhNWRiOWFhOWZiM2ZlOGRiN2Y3ZjMwNGFlZWRmOTg4NGRiYjZjZjI1ZGQ5NDNkZTQiLCJpYXQiOjE1OTcxMjk3MjAsIm5iZiI6MTU5NzEyOTcyMCwiZXhwIjoxNjI4NjY1NzE5LCJzdWIiOiI3Iiwic2NvcGVzIjpbXX0.PmPKlqkpQS0UxYVqXb6KdZtoouIUO7NyKW7bj9XweDMwmLUwmvvyCg7Z9wAMgJM0MsQl1-dTgd0XwP9e8Y9DbrAcvkjFVhOHnRPMsILSqNYk6XdFEIvqRtooOMjcdebRjuFO4Y39Tz8EAlfvVdYZu88J-j3ujJuIkE1fxo_wIeGL6gyjwUqeSiiUmM_BqFBvcBUfIaLYBXqkGtnrwNrKX10QcLfimCm6qgk3NtiQuEflFqWzkkq_uQoXOAOmk5UlPgUAWzxfvmmnSJ_B4rpKXa1DWCuP-ePL7ttA_DzuhWYlpTS4ovfdIvbTRg2p5eFOAqfko-1rzOwX1tjRQI4n6dsfVmyNAsfPTTVuF4se4gTomTx4uE-gDJ_yj52zeUHca5z4zkbTHZlH44TzF_MC2RfMQe6MLpNdT2EYlSzhWqlCDIMFF0-z101"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/requestBodies/response_403"
                    },
                    "422": {
                        "description": "Error: Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "description": "these are the fields of the request",
                                            "properties": {
                                                "email": {
                                                    "example": [
                                                        "El campo correo es obligatorio."
                                                    ]
                                                },
                                                "password": {
                                                    "example": [
                                                        "El campo password es obligatorio."
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "logout",
                "description": "",
                "operationId": "auth/logout",
                "responses": {
                    "200": {
                        "$ref": "#/components/requestBodies/response_200"
                    },
                    "401": {
                        "$ref": "#/components/requestBodies/response_401"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/auth/user/{profile_id}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticated user information",
                "description": "",
                "operationId": "/auth/user/{profile_id}",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/profile_id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "example": true
                                        },
                                        "message": {
                                            "example": "User login successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserAuthApiResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserAuthApiResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 161
                    },
                    "dni": {
                        "type": "string",
                        "example": "46485700M"
                    },
                    "name": {
                        "type": "string",
                        "example": "Carlos"
                    },
                    "lastname": {
                        "type": "string",
                        "example": "Perez Perez"
                    },
                    "fullName": {
                        "type": "string",
                        "example": "Carlos Perez Perez"
                    },
                    "email": {
                        "type": "string",
                        "example": "carlos@gmail.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+34622789562"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "profile_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "role": {
                        "type": "string",
                        "example": "admin"
                    }
                },
                "type": "object"
            }
        },
...
        "parameters": {
            "profile_id": {
                "name": "profile_id",
                "in": "path",
                "description": "Profile identifier",
                "required": true,
                "schema": {
                    "type": "integer",
                    "example": "11"
                }
            }
        },
        "requestBodies": {
            "response_200": {
                "description": "OK",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "success": {
                                    "example": true
                                },
                                "message": {
                                    "example": "Server message"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "response_401": {
                "description": "Error: Unauthorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "example": "Unauthenticated."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "response_403": {
                "description": "Error: Forbidden",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "success": {
                                    "example": false
                                },
                                "message": {
                                    "example": "You do not have permissions for the requested resources"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "Enter in the following field 'Bearer' followed by the token obtained in the endpoint /auth/login",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        []
    ]
}